fixup: fix all clippy suggestions and reorganize some code
