+++ docker-images/syntax-highlighter/src/bin/scip-snapshot.rs
@@ -1,6 +1,6 @@
 use std::fs;
 
-use sg_syntax::{dump_document, scip_index};
+use sg_syntax::{dump_document, treesitter_index};
 
 fn main() {
     if let Some(path) = std::env::args().nth(1) {
@@ -12,7 +12,7 @@ fn main() {
             }
         };
 
-        let document = match scip_index("go", &contents) {
+        let document = match treesitter_index("go", &contents) {
             Ok(document) => document,
             Err(err) => {
                 eprintln!("Failed to index document: {:?}", err);
